requirements.txt
selenium==4.18.1
pytest==8.0.2
pytest-html==4.1.1
allure-pytest==2.13.2
webdriver-manager==4.0.1
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.0.1
faker==22.6.0
jsonschema==4.21.1
requests==2.31.0 